Added AIHelper class
Clean up Faction enum, old Reputation Data
