+++ Assets/Scripts/EnemyAI/Faction2Script.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 
+using Model;
 using Model.AI;
 
 public class Faction2Script : AIShip
@@ -41,7 +42,7 @@ public class Faction2Script : AIShip
 
 
 		// If mission benefits our faction, fight harder
-		if (IsMissionAlly())
+		if (AIHelper.IsMissionAlly(faction))
 		{
 			switch (currentRole)
 			{
@@ -61,6 +62,8 @@ public class Faction2Script : AIShip
 
 	private void TankBehavior()
 	{
+		Transform player = GameObject.FindGameObjectWithTag("Player").transform;
+
 		// Tank: Stay in front of the player and absorb damage
 		Vector3 direction = (player.position - transform.position).normalized;
 		Vector3 avoidance = ComputeObstacleAvoidance(direction);
@@ -84,6 +87,8 @@ public class Faction2Script : AIShip
 
 	private void FastShipBehavior()
 	{
+		Transform player = GameObject.FindGameObjectWithTag("Player").transform;
+
 		// Fast Ship: Flank the player and evade attacks
 		Vector3 flankDirection = (player.position - transform.position).normalized;
 		flankDirection = Quaternion.Euler(0, 90, 0) * flankDirection; // Flank to the side
@@ -101,6 +106,8 @@ public class Faction2Script : AIShip
 
 	private void AttackerBehavior()
 	{
+		Transform player = GameObject.FindGameObjectWithTag("Player").transform;
+
 		// Attacker: Stay at a distance and deal damage
 		Vector3 direction = (player.position - transform.position).normalized;
 		Vector3 avoidance = ComputeObstacleAvoidance(direction);
