+++ Assets/Scripts/Manager/GameManager.cs
@@ -14,6 +14,8 @@ namespace Manager
     {
         public static GameManager Instance { get; private set; }
 
+        public PlayerReputation reputation = null;
+
         public GameObject playerPrefab;
         public GameObject playerShip;
 
@@ -58,6 +60,7 @@ namespace Manager
 		{
             this.isNewGame = gameData.isNewGame;
             this.hasPlayedTutorial = gameData.hasPlayedTutorial;
+            this.reputation = gameData.reputation;
         }
 
         // IDataPersistence
@@ -65,6 +68,7 @@ namespace Manager
 		{
             gameData.isNewGame = this.isNewGame;
             gameData.hasPlayedTutorial = this.hasPlayedTutorial;
+            gameData.reputation = this.reputation;
         }
 
         public void Play()
