+++ Assets/Scripts/Model/AI/BehaviorTree/AIShipBehaviorTree.cs
@@ -17,7 +17,7 @@ namespace Model.AI.BehaviorTree
 				new Sequence(new List<Node>
 				{
 					new Condition( () => aiShip.currentState != AIState.Seek ),
-					new Condition( () => aiShip.IsPlayerInRange() ),
+					new Condition( () => AIHelper.IsTargetInRange(aiShip) ),
 					new SetAIStateNode(aiShip, AIState.Seek)
 				}),
 				new SetAIStateNode(aiShip, AIState.Roam)
