+++ Assets/Scripts/Model/AIShip.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 536104ca9bb98fb4cbc90fd4c1de654c
+guid: 7c01cddc9d311ca41ac08a433c183545
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
