+++ Assets/Scripts/Model/JobSystem/Reputation.cs
@@ -1,15 +1,20 @@
-[System.Serializable]
-public class Reputation{
-    public RepType type;
-    public int value;
 
-    public Reputation(RepType _type, int _value)
+namespace Model
+{
+    [System.Serializable]
+    public class Reputation
     {
-        type = _type;
-        value = _value;
-    }
-    public void ChangeValue(int _value)
-    {
-        value += _value;
+        public RepType type;
+        public int value;
+
+        public Reputation(RepType _type, int _value)
+        {
+            type = _type;
+            value = _value;
+        }
+        public void ChangeValue(int _value)
+        {
+            value += _value;
+        }
     }
 }
