+++ Assets/Scripts/Model/Ship.cs
@@ -5,54 +5,54 @@ using Controller;
 namespace Model
 {
 
-	public class Ship : MonoBehaviour
-	{
-		public ShipData oriData;
+    public class Ship : MonoBehaviour
+    {
+        public ShipData oriData;
 
+        public int ammo;
+        public float health;
+        public float shields;
 
+        public virtual void SetStats()
+        {
+            health = oriData.maxHealth;
+            shields = oriData.maxShields;
+        }
 
-		public int ammo;
-		public float health;
-		public float shields;
+        public virtual bool IsJobTarget()
+        {
+            return false;
+        }
 
+        public virtual void ShieldRecover()
+        {
+            shields += oriData.shieldRegen;
+            shields = Mathf.Clamp(shields, 0, oriData.maxShields);
+        }
 
-		public virtual void SetStats()
-		{
-			health = oriData.maxHealth;
-			shields = oriData.maxShields;
-		}
+        public virtual bool TakeDamage(float _dmg)
+        {
+            health -= _dmg;
 
-		public virtual void ShieldRecover()
-		{
-			shields += oriData.shieldRegen;
-			shields = Mathf.Clamp(shields, 0, oriData.maxShields);
-		}
+            return CheckDestroyed();
+        }
 
+        public bool CheckDestroyed()
+        {
+            if (health > 0)
+                return false;
 
-		public virtual bool TakeDamage(float _dmg)
-		{
-			health -= _dmg;
+            JobController.Inst.OnObjDestroyed(transform.tag);
 
-			return CheckDestroyed();
-		}
+            this.gameObject.SetActive(false);
 
-		public bool CheckDestroyed()
-		{
-			if(health > 0)
-				return false;
+            return true;
+        }
 
-				
-			JobController.Inst.OnObjDestroyed(transform.tag);
-
-			this.gameObject.SetActive(false);
-
-			return true;	
-		}
-
-		public virtual void Leave()
-		{
-			JobController.Inst.OnObjLeave(transform.tag);
-			this.gameObject.SetActive(false);
-		}
-	}	
+        public virtual void Leave()
+        {
+            JobController.Inst.OnObjLeave(transform.tag);
+            this.gameObject.SetActive(false);
+        }
+    }
 }
